rename reducer files
