+++ src/redux/reducers/appsList.js
