+++ src/redux/reducers/deleteApp.js
