+++ src/redux/reducers/namespacesList.js
