+++ src/redux/reducers/storageClass.js
